# static/openapi.yaml
openapi: 3.0.3
info:
  title: TemplrPress API
  version: 0.4.0
  description: |
    HTTP surface of a running TemplrPress instance. The single binary serves
    both the embedded Next.js SPA (at `/`) and a small JSON API (under
    `/api/...`) used by the SPA and external integrators.
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:5000
...
paths:
...
  /api/openapi-spec:
    get:
      tags: [OpenAPI]
      summary: Fetch one OpenAPI spec
      parameters:
        - in: query
          name: spec
          required: false
          schema: { type: string }
          description: Spec name from `/api/config/openapi-specs`. Omit or use `_builtin` for the built-in spec.
      responses:
        "200":
          description: OpenAPI document (JSON)
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "404":
          description: Unknown spec name
...
  /api/cms/list:
    get:
      tags: [CMS]
      summary: List articles in a folder
      parameters:
        - in: query
          name: type
          schema: { type: string, default: docs, example: docs }
          description: Top-level folder under `content/` (e.g. `docs`, `blog`, `about`).
      responses:
        "200":
          description: Article list
          content:
            application/json:
              schema:
                type: object
                properties:
                  articles:
                    type: array
                    items: { $ref: "#/components/schemas/Frontmatter" }
...
components:
  schemas:
...
    Frontmatter:
      type: object
      properties:
        title:       { type: string }
        slug:        { type: string }
        description: { type: string }
        date:        { type: string, format: date-time }
        author:      { type: string }
        type:        { type: string }
        category:    { type: string }
        tags:        { type: array, items: { type: string } }
        coverImage:  { type: string }
        folder:      { type: string }
        url:         { type: string }
        sortOrder:   { type: integer }
...